% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satcat.r
\name{read_satcat}
\alias{read_satcat}
\title{Read in a complete SATCAT database complete with expanded status code, site code and
launch source codes.}
\usage{
read_satcat()
}
\description{
Read in a complete SATCAT database complete with expanded status code, site code and
launch source codes.
}
\examples{
satcat_df <- read_satcat()
}

