% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satcat-package.r
\docType{package}
\name{satcat}
\alias{satcat}
\alias{satcat-package}
\title{Tools to Work with the 'CelesTrak' Satellite Catalog API}
\description{
T.S. Kelso has been  transcribing NORAD's "resident space object" data for decades.
One component of of this is the 'SATCAT' satellite catalog (<https://celestrak.com/satcat/search.asp>),
which provides data on all known satellites launched since 1957.
}
\author{
Bob Rudis (bob@rud.is)
}

